use clang to compile android sample, fix sample compile before latest change

+++ cocos2dx/textures/CCTexture2D.cpp
@@ -805,7 +805,7 @@ void CCTexture2D::setAliasTexParameters()
 
     glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST );
 #if CC_ENABLE_CACHE_TEXTURE_DATA
-    ccTexParams texParams = {m_bHasMipmaps?GL_NEAREST_MIPMAP_NEAREST:GL_NEAREST,GL_NEAREST,GL_NONE,GL_NONE};
+    ccTexParams texParams = { m_bHasMipmaps ? (GLuint)GL_NEAREST_MIPMAP_NEAREST : (GLuint)GL_NEAREST, (GLuint)GL_NEAREST, (GLuint)GL_NONE, (GLuint)GL_NONE};
     VolatileTexture::setTexParameters(this, &texParams);
 #endif
 }
@@ -825,7 +825,7 @@ void CCTexture2D::setAntiAliasTexParameters()
 
     glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR );
 #if CC_ENABLE_CACHE_TEXTURE_DATA
-    ccTexParams texParams = {m_bHasMipmaps?GL_LINEAR_MIPMAP_NEAREST:GL_LINEAR,GL_LINEAR,GL_NONE,GL_NONE};
+    ccTexParams texParams = {m_bHasMipmaps ? (GLuint)GL_LINEAR_MIPMAP_NEAREST : (GLuint)GL_LINEAR, (GLuint)GL_LINEAR, (GLuint)GL_NONE, (GLuint)GL_NONE};
     VolatileTexture::setTexParameters(this, &texParams);
 #endif
 }
